added api calls to the data service
git-svn-id: file:///semplestb/final_repo@2597 02200ff9-b5b2-46f0-9171-221b09c08c7b

+++ SemplestWeb/SemplestAll/SharedResources/Services/ServiceClientWrapper.cs
@@ -302,7 +302,7 @@ namespace Semplest.SharedResources.Services
             return runBooleanMethod(ADENGINESERVICE, "scheduleChangePromotionStartDate", JsonConvert.SerializeObject(jsonHash));
         }
 
-        public bool scheduleUpdateBudget(int customerID, int promotionID, Double changeInBudget, List<String> adEngines)
+        public bool scheduleUpdateBudget(int customerID, int promotionID, Decimal changeInBudget, List<String> adEngines)
         {
             var jsonHash = new Dictionary<string, string>();
             jsonHash.Add("customerID", customerID.ToString());
